sistema-reportes-matriculas/server/requirements.txt
fastapi
uvicorn
python-multipart
easyocr
requests
python-dotenv